+++ util/msgpack/core_test.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with go-algorand.  If not, see <https://www.gnu.org/licenses/>.
 
-package core
+package msgpack
 
 import (
 	"encoding/base32"
